move storages metrics
