+++ src/common/metrics/src/metrics/mod.rs
@@ -13,3 +13,4 @@
 // limitations under the License.
 
 pub mod http;
+pub mod storage;
