--- src/common/storage/src/metrics/common.rs
@@ -1,30 +1,0 @@
-// Copyright 2021 Datafuse Labs
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use common_metrics::register_counter;
-use common_metrics::Counter;
-use lazy_static::lazy_static;
-
-lazy_static! {
-    static ref OMIT_FILTER_ROWGROUPS: Counter = register_counter("omit_filter_rowgroups");
-    static ref OMIT_FILTER_ROWS: Counter = register_counter("omit_filter_rows");
-}
-
-pub fn metrics_inc_omit_filter_rowgroups(c: u64) {
-    OMIT_FILTER_ROWGROUPS.inc_by(c);
-}
-
-pub fn metrics_inc_omit_filter_rows(c: u64) {
-    OMIT_FILTER_ROWS.inc_by(c);
-}
