+++ src/query/service/src/pipelines/builders/builder_on_finished.rs
@@ -19,9 +19,8 @@ use common_base::runtime::GlobalIORuntime;
 use common_catalog::table_context::TableContext;
 use common_exception::Result;
 use common_meta_app::principal::StageInfo;
+use common_metrics::storage::*;
 use common_pipeline_core::Pipeline;
-use common_storage::metrics::copy::metrics_inc_copy_purge_files_cost_milliseconds;
-use common_storage::metrics::copy::metrics_inc_copy_purge_files_counter;
 use common_storage::StageFileInfo;
 use common_storages_fuse::io::Files;
 use common_storages_stage::StageTable;
