+++ src/query/service/src/pipelines/processors/transforms/processor_deduplicate_row_number.rs
@@ -21,17 +21,13 @@ use common_expression::types::NumberDataType;
 use common_expression::types::UInt64Type;
 use common_expression::DataBlock;
 use common_expression::FromData;
+use common_metrics::storage::*;
 use common_pipeline_core::pipe::PipeItem;
 use common_pipeline_core::processors::port::InputPort;
 use common_pipeline_core::processors::port::OutputPort;
 use common_pipeline_core::processors::processor::ProcessorPtr;
 use common_pipeline_transforms::processors::transforms::AsyncAccumulatingTransform;
 use common_pipeline_transforms::processors::transforms::AsyncAccumulatingTransformer;
-use common_storage::metrics::merge_into::merge_into_distributed_apply_row_number;
-use common_storage::metrics::merge_into::merge_into_distributed_deduplicate_row_number;
-use common_storage::metrics::merge_into::merge_into_distributed_empty_row_number;
-use common_storage::metrics::merge_into::merge_into_distributed_init_unique_number;
-use common_storage::metrics::merge_into::merge_into_distributed_new_set_len;
 use itertools::Itertools;
 use log::info;
