+++ src/query/service/src/pipelines/processors/transforms/processor_extract_hash_table_by_row_number.rs
@@ -25,17 +25,13 @@ use common_expression::DataBlock;
 use common_expression::DataField;
 use common_expression::Scalar;
 use common_expression::Value;
+use common_metrics::storage::*;
 use common_pipeline_core::pipe::PipeItem;
 use common_pipeline_core::processors::port::InputPort;
 use common_pipeline_core::processors::port::OutputPort;
 use common_pipeline_core::processors::processor::Event;
 use common_pipeline_core::processors::processor::ProcessorPtr;
 use common_pipeline_core::processors::Processor;
-use common_storage::metrics::merge_into::merge_into_distributed_hashtable_empty_block;
-use common_storage::metrics::merge_into::merge_into_distributed_hashtable_fetch_row_number;
-use common_storage::metrics::merge_into::merge_into_distributed_hashtable_push_empty_null_block;
-use common_storage::metrics::merge_into::merge_into_distributed_hashtable_push_null_block;
-use common_storage::metrics::merge_into::merge_into_distributed_hashtable_push_null_block_rows;
 
 use super::hash_join::HashJoinBuildState;
 use super::processor_deduplicate_row_number::get_row_number;
