+++ src/query/service/src/sessions/query_ctx.rs
@@ -58,11 +58,11 @@ use common_meta_app::principal::UserInfo;
 use common_meta_app::schema::CatalogInfo;
 use common_meta_app::schema::GetTableCopiedFileReq;
 use common_meta_app::schema::TableInfo;
+use common_metrics::storage::*;
 use common_pipeline_core::InputError;
 use common_settings::ChangeValue;
 use common_settings::Settings;
 use common_sql::IndexType;
-use common_storage::metrics::copy::metrics_inc_filter_out_copied_files_request_milliseconds;
 use common_storage::CopyStatus;
 use common_storage::DataOperator;
 use common_storage::FileStatus;
