+++ src/query/sql/Cargo.toml
@@ -28,6 +28,7 @@ common-functions = { path = "../functions" }
 common-license = { path = "../../common/license" }
 common-meta-app = { path = "../../meta/app" }
 common-meta-types = { path = "../../meta/types" }
+common-metrics = { path = "../../common/metrics" }
 
 common-pipeline-core = { path = "../pipeline/core" }
 common-pipeline-sources = { path = "../pipeline/sources" }
