+++ src/query/storages/fuse/src/operations/merge_into/mutator/matched_mutator.rs
@@ -36,12 +36,7 @@ use common_expression::BlockMetaInfoDowncast;
 use common_expression::Column;
 use common_expression::DataBlock;
 use common_expression::TableSchemaRef;
-use common_storage::metrics::merge_into::metrics_inc_merge_into_accumulate_milliseconds;
-use common_storage::metrics::merge_into::metrics_inc_merge_into_apply_milliseconds;
-use common_storage::metrics::merge_into::metrics_inc_merge_into_deleted_blocks_counter;
-use common_storage::metrics::merge_into::metrics_inc_merge_into_deleted_blocks_rows_counter;
-use common_storage::metrics::merge_into::metrics_inc_merge_into_replace_blocks_counter;
-use common_storage::metrics::merge_into::metrics_inc_merge_into_replace_blocks_rows_counter;
+use common_metrics::storage::*;
 use itertools::Itertools;
 use log::info;
 use opendal::Operator;
