+++ src/query/storages/fuse/src/operations/merge_into/processors/processor_merge_into_matched_and_split.rs
@@ -27,6 +27,7 @@ use common_expression::DataSchemaRef;
 use common_expression::FieldIndex;
 use common_expression::Value;
 use common_functions::BUILTIN_FUNCTIONS;
+use common_metrics::storage::*;
 use common_pipeline_core::pipe::PipeItem;
 use common_pipeline_core::processors::port::InputPort;
 use common_pipeline_core::processors::port::OutputPort;
@@ -35,9 +36,6 @@ use common_pipeline_core::processors::processor::ProcessorPtr;
 use common_pipeline_core::processors::Processor;
 use common_sql::evaluator::BlockOperator;
 use common_sql::executor::MatchExpr;
-use common_storage::metrics::merge_into::merge_into_matched_operation_milliseconds;
-use common_storage::metrics::merge_into::metrics_inc_merge_into_append_blocks_counter;
-use common_storage::metrics::merge_into::metrics_inc_merge_into_append_blocks_rows_counter;
 
 use crate::operations::common::MutationLogs;
 use crate::operations::merge_into::mutator::DeleteByExprMutator;
