+++ src/query/storages/fuse/src/operations/merge_into/processors/transform_add_rownumber_column.rs
@@ -24,12 +24,12 @@ use common_expression::BlockEntry;
 use common_expression::DataBlock;
 use common_expression::FromData;
 use common_expression::Value;
+use common_metrics::storage::*;
 use common_pipeline_core::processors::port::InputPort;
 use common_pipeline_core::processors::port::OutputPort;
 use common_pipeline_core::processors::processor::ProcessorPtr;
 use common_pipeline_transforms::processors::transforms::Transform;
 use common_pipeline_transforms::processors::transforms::Transformer;
-use common_storage::metrics::merge_into::merge_into_distributed_generate_row_numbers;
 
 const PREFIX_OFFSET: usize = 48;
