+++ src/query/storages/parquet/src/parquet2/parquet_reader/reader.rs
@@ -30,8 +30,7 @@ use common_expression::DataBlock;
 use common_expression::DataSchema;
 use common_expression::DataSchemaRef;
 use common_expression::FieldIndex;
-use common_storage::metrics::copy::metrics_inc_copy_read_part_cost_milliseconds;
-use common_storage::metrics::copy::metrics_inc_copy_read_size_bytes;
+use common_metrics::storage::*;
 use common_storage::ColumnNodes;
 use opendal::BlockingOperator;
 use opendal::Operator;
