+++ src/query/storages/parquet/src/parquet_rs/parquet_reader/reader/full_reader.rs
@@ -18,8 +18,8 @@ use arrow_schema::ArrowError;
 use bytes::Bytes;
 use common_exception::Result;
 use common_expression::DataBlock;
-use common_storage::metrics::common::metrics_inc_omit_filter_rowgroups;
-use common_storage::metrics::common::metrics_inc_omit_filter_rows;
+use common_metrics::storage::metrics_inc_omit_filter_rowgroups;
+use common_metrics::storage::metrics_inc_omit_filter_rows;
 use futures::StreamExt;
 use opendal::Operator;
 use opendal::Reader;
